requirements.txt
configparser==4.0.2
contextlib2==0.6.0.post1
dataclasses==0.6
importlib-metadata==2.1.3
pathlib2==2.3.7.post1
scandir==1.10.0
six==1.16.0
SQLAlchemy==1.4.49
typing==3.10.0.0
zipp==1.2.0
configparser==4.0.2
contextlib2==0.6.0.post1
dataclasses==0.6
importlib-metadata==2.1.3
pathlib2==2.3.7.post1
scandir==1.10.0
six==1.16.0
SQLAlchemy==1.4.49
typing==3.10.0.0
zipp==1.2.0
configparser==4.0.2
contextlib2==0.6.0.post1
dataclasses==0.6
importlib-metadata==2.1.3
pathlib2==2.3.7.post1
psycopg2-binary==2.8.6
python-dotenv==0.18.0
scandir==1.10.0
six==1.16.0
SQLAlchemy==1.4.49
typing==3.10.0.0
zipp==1.2.0
